Update api-management-virtual-network-troubleshooting.md
I have come across a scenario where we were unable to revert back the APIM instance to its previous subnet. 

I have added a couple of scenarios/ challenges which could help customers resolve the issue themselves. 

I am from the APIM Support team and these are some of the solutions we suggested our customers to follow in order to achieve this goal.